整理代码，激活项目。
