+++ .gitignore
@@ -2,3 +2,4 @@ CMakeFiles
 build
 .vscode
 .history
+.project
