+++ Main.h
@@ -5,6 +5,7 @@
 #include <vector>
 #include "ChanlunZb.h"
 #include "ChanlunXg.h"
+#include "BaoHan.h"
 #include "Bi.h"
 #include "Duan.h"
 #include "ZhongShu.h"
