+++ README.md
@@ -1,15 +1,35 @@
-### Г╪√Х╞▒И║╧Г⌡╝
+### хГ╨н╠ЮрК
 
-Е°╗И║╧Г⌡╝Ф═╧Г⌡╝Е╫∙Д╦▀Ф┴╖Х║▄Д╩╔Д╦▀Е▒╫Д╩╓
+#### г╟лАлУ╪Ч
+
+- Visual Studio 2015 ╩Руъ Visual Studio 2017 ╩Руъж╩р╙ Windows Build Toolsр╡йг©ирт╠ЮрК╣дё╛фДкЭ╠ЮрК╧╓╬ъц╩сп╡Бйт╣╚йгр╡с╕╦ц©ирт╠ЮрК║ё
+- ╟╡в╟CMake║ё
+
+### ╠ЮрКцЭаН
+
+тзоНд©╣д╦Ыд©б╪обрю╢нж╢ппртобцЭаН║ё
 
 ```
 mkdir build
 cd build
 cmake ..
 cmake --build . --config Release
+```
+
+╩РуъсцобцФцЭаНиЗЁиDebug╟Ф╠╬║ё
+
+```
 cmake --build . --config Debug
 ```
 
-### Д╩≤Х╢╧Ф■╞Ф▄│
+### ╫╩аВ
+- н╒пе╧╚жз╨е: sokldcty
 - WeChat: kldcty
 - QQ: 1106628276
+
+н╒пех╨гК╪сн╒пе╩РуъQQвия╞ё╛н╒пех╨╣д╧ФтРйгуБяЫ╣д
+- иЙгК╪сх╨уъпХр╙╥╒р╩╦Ж╨Л╟Э╦ЬфДкШх╨Ёит╠гю╨Л╟Эё╛жасзн╙й╡ц╢р╙╥╒╨Л╟ЭдгйгиХжцр╩╦Ж╪схК╣дце╪Вё╛оёмШ╢С╪рхКх╨йгю╢уФпд╫╩аВ╣д║ё
+- н╒пех╨юО©ирт╫╩аВ╡Ьбш╪╪йУ╪╟╧иф╠╫╩рвоЮ╧ь╩╟лБ║ё
+- ╣╠х╩ох╪схК╣дм╛я╖©иртгю╣╫╨С╪схКм╛я╖╣д╨Л╟Э║ё
+
+уБ╦ЖоНд©ря╬╜╩д╥ор╩╤нй╠╪Дакё╛оёмШсцуБяЫ╣д╥╫й╫жьпб╪╓╩НуБ╦ЖоНд©║ё
